Fixed same ID issue
A test patch to fix the issue where devices with the same ID as the keydial are selected before the keydial. If the code is ran as root, this process is automatic but if it is ran as a user, you will be prompted to select your device before the driver can run.